requirements.txt
websocket-client
python-dotenv
line-bot-sdk
tornado
flask